Feature.test end to end

+++ src/main/kotlin/ar/edu/unq/desapp/grupoA/CryptoExchangeApi/model/Active.kt
@@ -2,7 +2,7 @@ package ar.edu.unq.desapp.grupoA.CryptoExchangeApi.model
 
 class Active(
     val crypto: String,
-    val criptoNominalQuantity: Float,
+    val criptoNominalQuantity: Double,
     val usdPrice: Float,
     val argPrice: Float,
 ) {}
